SendCryptoDetailsView refactored
