+++ VultisigApp/VultisigApp/Utils/KeyboardObserver.swift
@@ -8,7 +8,6 @@
 import SwiftUI
 import Combine
 
-#if os(iOS)
 final class KeyboardObserver: ObservableObject {
     
     @MainActor @Published var keyboardHeight: CGFloat = 0
@@ -16,6 +15,7 @@ final class KeyboardObserver: ObservableObject {
     private var cancellables: Set<AnyCancellable> = []
 
     init() {
+#if os(iOS)
         NotificationCenter.default.publisher(for: UIResponder.keyboardWillShowNotification)
             .compactMap { $0.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect }
             .map { $0.height }
@@ -26,6 +26,6 @@ final class KeyboardObserver: ObservableObject {
             .map { _ in CGFloat(0) }
             .assign(to: \.keyboardHeight, on: self)
             .store(in: &cancellables)
+#endif
     }
 }
-#endif
